shift everything related to overset grids to separate overset_grids directory
