+++ include/exadg/poisson/overset_grids/user_interface/declare_get_application.h
@@ -19,8 +19,8 @@
  *  ______________________________________________________________________
  */
 
-#ifndef INCLUDE_EXADG_POISSON_USER_INTERFACE_DECLARE_GET_APPLICATION_H_
-#define INCLUDE_EXADG_POISSON_USER_INTERFACE_DECLARE_GET_APPLICATION_H_
+#ifndef INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_DECLARE_GET_APPLICATION_H_
+#define INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_DECLARE_GET_APPLICATION_H_
 
 #include <exadg/poisson/user_interface/application_base.h>
 
@@ -29,12 +29,11 @@ namespace ExaDG
 namespace Poisson
 {
 template<int dim, int n_components, typename Number>
-std::shared_ptr<ApplicationBase<dim, n_components, Number>>
-get_application(std::string input_file, MPI_Comm const & comm);
+std::shared_ptr<ApplicationOversetGridsBase<dim, n_components, Number>>
+get_application_overset_grids(std::string input_file, MPI_Comm const & comm);
 
 } // namespace Poisson
 } // namespace ExaDG
 
 
-
-#endif /* INCLUDE_EXADG_POISSON_USER_INTERFACE_DECLARE_GET_APPLICATION_H_ */
+#endif /* INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_DECLARE_GET_APPLICATION_H_ */
