+++ include/exadg/poisson/overset_grids/user_interface/implement_get_application.h
@@ -19,8 +19,8 @@
  *  ______________________________________________________________________
  */
 
-#ifndef INCLUDE_EXADG_POISSON_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_OVERSET_GRIDS_H_
-#define INCLUDE_EXADG_POISSON_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_OVERSET_GRIDS_H_
+#ifndef INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_H_
+#define INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_H_
 
 #include <exadg/poisson/user_interface/application_base.h>
 
@@ -39,4 +39,4 @@ get_application_overset_grids(std::string input_file, MPI_Comm const & comm)
 } // namespace ExaDG
 
 
-#endif /* INCLUDE_EXADG_POISSON_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_OVERSET_GRIDS_H_ */
+#endif /* INCLUDE_EXADG_POISSON_OVERSET_GRIDS_USER_INTERFACE_IMPLEMENT_GET_APPLICATION_H_ */
