+++ include/exadg/poisson/solver_poisson.h
@@ -24,8 +24,8 @@
 
 // ExaDG
 #include <exadg/matrix_free/matrix_free_data.h>
+#include <exadg/poisson/overset_grids/user_interface/application_base.h>
 #include <exadg/poisson/spatial_discretization/operator.h>
-#include <exadg/poisson/user_interface/application_base.h>
 
 namespace ExaDG
 {
